fixed Multicall contract and tests. improved mullticall in ArgentAccount. introduction of utilities file for tests.
