+++ test/argent_account.py
@@ -8,7 +8,7 @@ from starkware.starknet.business_logic.state import BlockInfo
 from starkware.starknet.business_logic.transaction_execution_objects import Event
 from starkware.starknet.compiler.compile import get_selector_from_name
 from utils.Signer import Signer
-from utils.deploy import deploy
+from utils.utilities import deploy, assert_revert
 from utils.TransactionSender import TransactionSender
 
 LOGGER = logging.getLogger(__name__)
@@ -26,15 +26,6 @@ VERSION = 206933470768 # '0.2.0' = 30 2E 32 2E 30 = 0x302E322E30 = 206933470768
 
 ESCAPE_TYPE_GUARDIAN = 0
 ESCAPE_TYPE_SIGNER = 1
-
-
-async def assert_revert(expression):
-    try:
-        await expression
-        assert False
-    except StarkException as err:
-        _, error = err.args
-        assert error['code'] == StarknetErrorCode.TRANSACTION_FAILED
 
 def assert_event_emmited(tx_exec_info, from_address, name, data):
     assert Event(
