+++ test/multicall.py
@@ -2,13 +2,10 @@ import pytest
 import asyncio
 from starkware.starknet.testing.starknet import Starknet
 from starkware.starknet.public.abi import get_selector_from_name
-from utils.deploy import deploy
+from utils.utilities import deploy, str_to_felt, uint
 
 user1 = 0x69221ff9023c4d7ba9123f0f9c32634c23fc5776d86657f464ecb51fd811445
 user2 = 0x72648c3b1953572d2c4395a610f18b83cca14fa4d1ba10fc4484431fd463e5c
-
-def uint(a):
-    return(a, 0)
 
 @pytest.fixture(scope='module')
 def event_loop():
@@ -23,18 +20,15 @@ async def get_starknet():
 async def test_multicall(get_starknet):
     starknet = get_starknet
     multicall = await deploy(starknet, "contracts/Multicall.cairo")
-    erc20_1 = await deploy(starknet, "contracts/ERC20.cairo")
-    erc20_2 = await deploy(starknet, "contracts/ERC20.cairo")
-
-    await erc20_1.mint(user1, uint(100)).invoke()
-    await erc20_2.mint(user2, uint(200)).invoke()
+    erc20_1 = await deploy(starknet, "contracts/lib/ERC20.cairo", [str_to_felt('token1'), str_to_felt('T1'), user1])
+    erc20_2 = await deploy(starknet, "contracts/lib/ERC20.cairo", [str_to_felt('token2'), str_to_felt('T2'), user2])
 
     response = await multicall.multicall([
-        erc20_1.contract_address, get_selector_from_name('get_decimals'), 0,
-        erc20_1.contract_address, get_selector_from_name('balance_of'), 1, user1,
-        erc20_2.contract_address, get_selector_from_name('balance_of'), 1, user2
+        erc20_1.contract_address, get_selector_from_name('decimals'), 0,
+        erc20_1.contract_address, get_selector_from_name('balanceOf'), 1, user1,
+        erc20_2.contract_address, get_selector_from_name('balanceOf'), 1, user2
     ]).call()
     print(response.result)
     assert response.result.result[0] == 18
-    assert response.result.result[1] == 100
-    assert response.result.result[3] == 200
+    assert response.result.result[1] == 1000
+    assert response.result.result[3] == 1000
